[FIX] Do not hide lots when active_test is false
There was a problem in the way we were hidding lots from the users: when users create prodlots, the client calls search(... [('id','=',<id of the new record>)]. Since we were hiding the lots as soon as the lock checkbox was checked, search() returned False which the client didn't expect.
Fixed by letting search return locked lots as the context also asks it to return the inactive lots, which the client does when new records are created.