[DEL] drop obsolete modules
stock_disable_backorder was a workaround for the crappy partial transfer wizard in v7. The one in v8 does not suffer from the same problems so we don't need that anymore.
stock_move_initial_qty was useful because we didn't have a trace of what the initial order was, but now that we have stock moves & pack operations we don't need it anymore